+++ rollup.config.mjs
@@ -1,22 +1,26 @@
-import typescript from "rollup-plugin-typescript2"
+import typescript from "@rollup/plugin-typescript"
 import { nodeResolve } from "@rollup/plugin-node-resolve"
 import terser from "@rollup/plugin-terser"
 
 export default {
   input: "./src/index.ts",
-  output: {
-    file: "lib/index.js",
-    format: "cjs",
-    exports: "auto",
-  },
+  output: [
+    {
+      file: "lib/index.cjs",
+      format: "cjs",
+      exports: "auto",
+    },
+    {
+      file: "lib/index.mjs",
+      format: "es",
+    },
+  ],
   plugins: [
     typescript({
-      direction: true,
       removeComments: true,
-      // 使用声明生成路径配置
-      useTsconfigDeclarationDir: true,
+      tsconfig: "./tsconfig.rollup.json",
     }),
     nodeResolve(),
-    terser()
+    terser(),
   ],
 }